Fixed an ecmdquery chips print bug
There was an issue when going from non thread cores to thread
cores that led to improper printing

The old code was one big block that did both the thread finding and
the output printing.

To fix the issue, I broke it into two loops.  The first does the
thread finding, the second does the printing.

Signed-off-by: Jason Albert <albertj@us.ibm.com>